ssh-login-brute-forcing/requirements.txt
paramiko==2.12.0
pwntools==4.8.0